Move writeZone to a separate util dir
The problem with `writeZone` is that it produces a derivation by using
`writeTextFile` from nixpkgs, which, in turn, is tricky to access from
within a flake, since it depends on the `system` attribute.

Moving it to a separate module will allows us to clearly separate “pure”
and “impure” functions and expose them form the flake accordingly.