Pass only `lib` to pure parts of the code
Previously, we were passing `pkgs` to all the parts of the code that
generate the records, but all of them were only using `pkgs.lib`.

Now we only pass `lib`, since it is “pure”, i.e. does not depend on the
stdenv and thus the `system` we are building on/for.

This will allow us to correctly expose `writeZone` through the flake
interface.