Dual-license everything as MPL-2.0 or MIT
MIT is a pretty bad license, and I have chosed it only to be compatible
with nixpkgs, in case this work will get merged there in the future.

When copyright is taken seriously, dealing with MIT is hard, so, I
think, some people might find it easier to use this code if it will be
license as MPL-2.0.

Now I don’t think that this will ever get merged into nixpkgs and I
don’t think it even makes sense to do so, but, I guess, it does not hurt
to keep the MIT option just in case.