DMARC: Fix the order of `p`
The specfication could be more clear, but after carefully reading it
I realised that not only `v=` MUST be the first tag, it also MUST
be followed by `p=`.

Tweak the DMARC record generation code to treat `p=` specially by
pulling it in the front of the list of tags. It is a required tag
anyway, so it is pretty straightforward.